(PC-23646)[PRO] refactor: test the VenueCreationLinks component directly
